Add reset to defaults functionality
