+++ Launcher/Classes/ViewControllers/TasksViewController.swift
@@ -64,10 +64,6 @@ class TasksViewController: NSViewController {
         }
         
         killProcessScreenshot()
-        // removes NSUserDefaults for test
-//        if let appDomain = Bundle.main.bundleIdentifier {
-//            UserDefaults.standard.removePersistentDomain(forName: appDomain)
-//        }
 
         textView.backgroundColor = .darkAquamarine
         textView.textColor = .white
